# From string validation