Resolves #71.

---

commit 06d30850b04b548f23e58d1d3ed991452f1338a6 (HEAD -> from-string, smyrman/from-string)
Author: Sindre Myren <sindre@clarify.io>
Date:   Thu Mar 10 19:39:00 2022 +0100

    fix: let decode look for additional base32 padding

    Update FromString and XID.TextUnmarshal so that it looks for discarded
    bits in the final source character. This ensures that XIDs that have
    been manually tampered with in a way that's ignored by base32 decode,
    will not pass as valid.

commit f28827202bb2c353a9f77d4d4b6a6a0a7e6b001a
Author: Sindre Myren <sindre@clarify.io>
Date:   Fri Mar 4 09:13:06 2022 +0100

    add benchmark and new failing test for FromString

commit 2171fc4df3cfc3d83ab324dd8d0ce1acf9d2b326
Author: Sindre Myren <sindre@clarify.io>
Date:   Fri Mar 4 09:12:03 2022 +0100

    error.go: make ErrInvalidID a const

    Make ErrInvalidID a constant instead of a variable. This prevent it from
    being changed by external packages; a behavior that although allowed by
    the compiler, should probably be considered an invalid operation.

commit b64cbc7cd2419bc72233b1de6286c91a92ebf6f4
Author: Sindre Myren <sindre@clarify.io>
Date:   Fri Mar 4 09:09:13 2022 +0100

    chore: address linter issues

    Perform changes suggested by linter; no functional changes.