@smyrman 2022-03-04
```txt
go test -bench 'FromString*'                                                                                                                                                                                                                                                                                           :(
goos: darwin
goarch: amd64
pkg: github.com/rs/xid
cpu: Intel(R) Core(TM) i7-6567U CPU @ 3.30GHz
BenchmarkFromString-4    	83381887	        13.95 ns/op
BenchmarkFromString2-4   	57998780	        19.95 ns/op
PASS
ok  	github.com/rs/xid	4.594s
go test -bench 'FromString*'  16.73s user 0.52s system 278% cpu 6.191 total
```