@smyrman 2022-03-04
```
% go test -bench 'FromString*' -benchmem
goos: darwin
goarch: amd64
pkg: github.com/rs/xid
cpu: Intel(R) Core(TM) i7-6567U CPU @ 3.30GHz
BenchmarkFromString-4           76597575                14.00 ns/op            0 B/op          0 allocs/op
BenchmarkFromString2-4          59851599                20.47 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/rs/xid       3.378s
go test -bench 'FromString*' -benchmem  12.67s user 0.83s system 214% cpu 6.290 total
```