@smyrman 2022-03-04
Verifying that `fromString2` actually solves #71:

```
go test -v -run 'TestFromString.*Quick.*'
=== RUN   TestFromStringQuick
    id_test.go:292: comparing XIDs:
        a: "c8gtev2sahsk17no3shg"
        b: "c8gtev2sahsk17no3shi" (index 19 changed to i)
    id_test.go:307: #1: failed on input xid.ID{0x62, 0x21, 0xd7, 0x7c, 0x5c, 0x54, 0x79, 0x40, 0x9e, 0xf8, 0x1f, 0x23}, 0x69
--- FAIL: TestFromStringQuick (0.00s)
=== RUN   TestFromString2Quick
--- PASS: TestFromString2Quick (0.00s)
=== RUN   TestFromString2QuickInvalidChars
--- PASS: TestFromString2QuickInvalidChars (0.00s)
FAIL
exit status 1
FAIL	github.com/rs/xid	0.293s
1 smyrman@patchi ~/Code/xid (git)-[from-strin
```