@smyrman 2022-03-04
I will clean up this PR if we decide that we want the semantics of `fromString2` over `FromString`, and that we are _happy_ with the performance implications.