@smyrman 2022-03-10
Updated benchmarks.

# Before change
```
$ go test -bench '.*' -run '^$'
goos: darwin
goarch: amd64
pkg: github.com/rs/xid
cpu: Intel(R) Core(TM) i7-6567U CPU @ 3.30GHz
BenchmarkNew-4          	29175643	        40.74 ns/op
BenchmarkNewString-4    	19683991	        54.16 ns/op
BenchmarkFromString-4   	77936979	        14.54 ns/op
PASS
ok  	github.com/rs/xid	5.598s
go test -bench '.*' -run '^$'  20.10s user 0.54s system 320% cpu 6.441 total
```
New failing test:
```
$ go test
--- FAIL: TestFromStringQuick (0.00s)
    id_test.go:265: comparing XIDs:
        a: "c8l4djisahsnf5k88go0"
        b: "c8l4djisahsnf5k88goa" (index 19 changed to a)
    id_test.go:280: #1: failed on input xid.ID{0x62, 0x2a, 0x46, 0xce, 0x5c, 0x54, 0x79, 0x77, 0x96, 0x88, 0x44, 0x30}, 0x61
--- FAIL: TestFromStringQuickInvalidChars (0.00s)
    id_test.go:292: comparing XIDs:
        a: "c8l4djisahsnf5k88gvg"
        b: "c8l4djisahsnf5k88gvl" (index 19 changed to l)
    id_test.go:307: #15: failed on input xid.ID{0x62, 0x2a, 0x46, 0xce, 0x5c, 0x54, 0x79, 0x77, 0x96, 0x88, 0x44, 0x3f}, 0x6c
FAIL
exit status 1
FAIL	github.com/rs/xid	0.160s
```

# After change
```
$ go test -bench '.*' -run '^$'
goos: darwin
goarch: amd64
pkg: github.com/rs/xid
cpu: Intel(R) Core(TM) i7-6567U CPU @ 3.30GHz
BenchmarkNew-4          	28826906	        40.04 ns/op
BenchmarkNewString-4    	20088812	        50.65 ns/op
BenchmarkFromString-4   	64955624	        17.20 ns/op
PASS
ok  	github.com/rs/xid	5.484s
go test -bench '.*' -run '^$'  19.65s user 0.95s system 237% cpu 8.678 total
```

Passing tests:
```
$ go test
PASS
ok  	github.com/rs/xid	0.279s
```