@smyrman 2022-03-10
> Do you need to do something more before merge?

Well, I have some questions.

1. I changed errors to be `const` -- this _might_ be viewed as a breaking change. Is this a change that you want, or should I pull it out?

2. Do you want a separate error for invalid padding, or should there just be one error for all invalid IDs?