@rs 2022-03-10
Const is fine and same error may be fine as well, I let you judge. I'm not sure a separate error is actionable.